implemented video sample meta init part

+++ include/VideoSampleInfo.h
@@ -14,13 +14,13 @@ public:
     int getFlag() const;
 
     void setSize(int inputSize);
-    void setOffset(int inputOffset);
+    void setOffset(int64_t inputOffset);
     void setFlag(int inputFlag);
 
 private:
     std::vector<RtpMetaInfo> rtpMetaInfos;
     int size;
-    int offset;
+    int64_t offset;
     int flag; // 1: Key(== I) frame, 0: P frame
 };
